pyproject.toml
[tool.poetry]
name = "urlgenerator"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
